+++ routes/web.php
@@ -29,37 +29,40 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', [ListingController::class, 'index']);
 
 // Show Create form
-Route::get('/listing/create', [ListingController::class, 'create']);
+Route::get('/listing/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store Listing Data
-Route::post('/listing', [ListingController::class, 'store']);
+Route::post('/listing', [ListingController::class, 'store'])->middleware('auth');
 
 // Edit Listing Data
-Route::get('/listing/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listing/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Update Listing Data
-Route::put('/listing/{listing}', [ListingController::class, 'update']);
+Route::put('/listing/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete Listing Data
-Route::delete('/listing/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listing/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // Single Listing
 Route::get('/listing/{listing}', [ListingController::class, 'show']);
 
+// Mange Listing
+Route::get('/listings/manage', [ListingController::class, 'manage']);
+
 // User Registration
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // Store User
 Route::post('/users', [UserController::class, 'store']);
 
 // Logout User
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 // Login User
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Login
-Route::post('//users/auth', [UserController::class, 'authenticate']);
+Route::post('/users/auth', [UserController::class, 'authenticate']);
